+++ src/lib/config/ace/src/config.cpp
@@ -11,6 +11,7 @@
 #include "ace/moreutils.hpp"
 
 #include "etl/string.h"
+#include "etl/algorithm.h"
 
 #include "pico/rand.h"
 
@@ -74,7 +75,7 @@ GATAS::PostConstruct Config::postConstruct()
         else
         {
             statistics.location = PERSISTENT;
-            statistics.persistentStoreSize = static_cast<uint16_t>(strlen((const char *)permanentStore.data()) + 1);
+            statistics.persistentStoreSize = strlen((const char *)permanentStore.data()) + 1;
             serializeToVolatile();
         }
     }
@@ -252,7 +253,7 @@ void Config::serializeToPersistent()
 
     permanentStore.rewind();
     permanentStore.write(volatileStore.data(), strlen((const char *)volatileStore.data()) + 1);
-    statistics.persistentStoreSize = static_cast<uint16_t>(strlen((char *)permanentStore.data()) + 1);
+    statistics.persistentStoreSize = strlen((char *)permanentStore.data()) + 1;
 }
 
 bool Config::deleteData(const etl::string_view fullPath)
